refactor(tests): make sync and async connection tests more similar
